debug halve

+++ csf_assign01/fixedpoint.c
@@ -199,7 +199,7 @@ Fixedpoint fixedpoint_halve(Fixedpoint val) {
   val.fractional = (val.fractional >> 1); //divide frac by two
   if ((val.whole & 1) == 1) { //need to shift 1 to frac
     // val.fractional |= (1<<63);
-    val.fractional += 18446744073709551616;
+    val.fractional += (uint64_t) 18446744073709551616;
   }
   val.whole = (val.whole >> 1); //divide whole by two
   return val;
